+++ core/products/future_test.go
@@ -27,18 +27,18 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestScalingOfSettlementPrice(t *testing.T) {
-	t.Run("No scaling needed for settlement price for asset decimals", testNoScalingNeeded)
-	t.Run("Need to scale up the settlement price for asset decimals", testScalingUpNeeded)
-	t.Run("Need to scale down the settlement price for asset decimals no loss of precision", testScalingDownNeeded)
-	t.Run("Need to scale down the settlement price for asset decimals with loss of precision", testScalingDownNeededWithPrecisionLoss)
+func TestScalingOfSettlementData(t *testing.T) {
+	t.Run("No scaling needed for settlement data for asset decimals", testNoScalingNeeded)
+	t.Run("Need to scale up the settlement data for asset decimals", testScalingUpNeeded)
+	t.Run("Need to scale down the settlement data for asset decimals no loss of precision", testScalingDownNeeded)
+	t.Run("Need to scale down the settlement data for asset decimals with loss of precision", testScalingDownNeededWithPrecisionLoss)
 }
 
 func testNoScalingNeeded(t *testing.T) {
 	ft := testFuture(t)
 
-	// settlement price is in 5 decimal places, asset in 5 decimal places => no scaling
-	scaled, err := ft.future.ScaleSettlementPriceToDecimalPlaces(num.NewUint(100000), 5)
+	// settlement data is in 5 decimal places, asset in 5 decimal places => no scaling
+	scaled, err := ft.future.ScaleSettlementDataToDecimalPlaces(num.NewUint(100000), 5)
 	require.NoError(t, err)
 	require.Equal(t, num.NewUint(100000), scaled)
 }
@@ -46,8 +46,8 @@ func testNoScalingNeeded(t *testing.T) {
 func testScalingUpNeeded(t *testing.T) {
 	ft := testFuture(t)
 
-	// settlement price is in 5 decimal places, asset in 10 decimal places => x10^5
-	scaled, err := ft.future.ScaleSettlementPriceToDecimalPlaces(num.NewUint(100000), 10)
+	// settlement data is in 5 decimal places, asset in 10 decimal places => x10^5
+	scaled, err := ft.future.ScaleSettlementDataToDecimalPlaces(num.NewUint(100000), 10)
 	require.NoError(t, err)
 	require.Equal(t, num.NewUint(10000000000), scaled)
 }
@@ -55,8 +55,8 @@ func testScalingUpNeeded(t *testing.T) {
 func testScalingDownNeeded(t *testing.T) {
 	ft := testFuture(t)
 
-	// settlement price is in 5 decimal places, asset in 3 decimal places => x10^-2
-	scaled, err := ft.future.ScaleSettlementPriceToDecimalPlaces(num.NewUint(100000), 3)
+	// settlement data is in 5 decimal places, asset in 3 decimal places => x10^-2
+	scaled, err := ft.future.ScaleSettlementDataToDecimalPlaces(num.NewUint(100000), 3)
 	require.NoError(t, err)
 	require.Equal(t, num.NewUint(1000), scaled)
 }
@@ -64,8 +64,8 @@ func testScalingDownNeeded(t *testing.T) {
 func testScalingDownNeededWithPrecisionLoss(t *testing.T) {
 	ft := testFuture(t)
 
-	// settlement price is in 5 decimal places, asset in 3 decimal places => x10^-2
-	scaled, err := ft.future.ScaleSettlementPriceToDecimalPlaces(num.NewUint(123456), 3)
+	// settlement data is in 5 decimal places, asset in 3 decimal places => x10^-2
+	scaled, err := ft.future.ScaleSettlementDataToDecimalPlaces(num.NewUint(123456), 3)
 	require.NoError(t, err)
 	require.Equal(t, num.NewUint(1234), scaled)
 }
@@ -85,7 +85,7 @@ func testFuture(t *testing.T) *tstFuture {
 	f := &types.Future{
 		SettlementAsset: "ETH",
 		QuoteName:       "ETH",
-		OracleSpecForSettlementPrice: &types.OracleSpec{
+		OracleSpecForSettlementData: &types.OracleSpec{
 			PubKeys: []string{"0xDEADBEEF"},
 			Filters: []*types.OracleSpecFilter{
 				{
@@ -110,7 +110,7 @@ func testFuture(t *testing.T) *tstFuture {
 			},
 		},
 		OracleSpecBinding: &types.OracleSpecBindingForFuture{
-			SettlementPriceProperty:    "price.ETH.value",
+			SettlementDataProperty:     "price.ETH.value",
 			TradingTerminationProperty: "trading.termination",
 		},
 		SettlementDataDecimals: 5,