+++ datanode/gateway/graphql/resolvers_test.go
@@ -93,7 +93,7 @@ func getTestMarket() *types.Market {
 				Product: &types.Instrument_Future{
 					Future: &types.Future{
 						SettlementAsset: "Ethereum/Ether",
-						OracleSpecForSettlementPrice: &oraclesv1.OracleSpec{
+						OracleSpecForSettlementData: &oraclesv1.OracleSpec{
 							PubKeys: []string{"0xDEADBEEF"},
 							Filters: []*oraclesv1.Filter{
 								{
@@ -118,7 +118,7 @@ func getTestMarket() *types.Market {
 							},
 						},
 						OracleSpecBinding: &types.OracleSpecToFutureBinding{
-							SettlementPriceProperty:    "prices.ETH.value",
+							SettlementDataProperty:     "prices.ETH.value",
 							TradingTerminationProperty: "trading.terminated",
 						},
 					},