+++ datanode/sqlsubscribers/market_created_test.go
@@ -56,7 +56,7 @@ func getTestMarket() types.Market {
 					Future: &types.Future{
 						SettlementAsset: "",
 						QuoteName:       "",
-						OracleSpecForSettlementPrice: &types.OracleSpec{
+						OracleSpecForSettlementData: &types.OracleSpec{
 							ID:        "",
 							CreatedAt: 0,
 							UpdatedAt: 0,
@@ -73,7 +73,7 @@ func getTestMarket() types.Market {
 							Status:    0,
 						},
 						OracleSpecBinding: &types.OracleSpecBindingForFuture{
-							SettlementPriceProperty:    "",
+							SettlementDataProperty:     "",
 							TradingTerminationProperty: "",
 						},
 					},